+++ packages/core/tests/fixtures/__snapshots__/docs/grid/alignment.react.tsx
@@ -5,7 +5,6 @@ export const WiremdComponent: React.FC = () => {
     <div className="wmd-root">
   <div className="wmd-grid wmd-grid-2" style={{ '--grid-columns': 2 } as React.CSSProperties}>
     <div className="wmd-grid-item">
-      <h3 className="wmd-h3"></h3>
       <div className="wmd-row">
         <div className="wmd-grid-item">
           <button className="wmd-button">
@@ -18,7 +17,6 @@ export const WiremdComponent: React.FC = () => {
       </div>
     </div>
     <div className="wmd-grid-item wmd-align-right">
-      <h3 className="wmd-h3 wmd-right"></h3>
       <button className="wmd-button wmd-button-primary">
         + New Sprint
       </button>